staticdata: remove `reinit_ccallable`
If JuliaLang/julia#56499 is merged, the functionality of
`jl_reinit_ccallable` will become incomplete. However, it is
questionable whether `jl_reinit_ccallable` is truly necessary.
It seems to have been added in JuliaLang/julia#44527, but all test cases
appear to pass without it. Therefore, it might be safe to remove it
altogether.